Move core::iter adapters to adapters.rs
